SP-359 refactor: repository 디렉터리 구조 변경
