+++ src/main/java/com/ludo/study/studymatchingplatform/study/repository/study/category/CategoryJpaRepository.java
@@ -2,7 +2,7 @@ package com.ludo.study.studymatchingplatform.study.repository.study.category;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ludo.study.studymatchingplatform.study.domain.study.Category;
+import com.ludo.study.studymatchingplatform.study.domain.study.category.Category;
 
 public interface CategoryJpaRepository extends JpaRepository<Category, Long> {
