+++ src/main/java/com/ludo/study/studymatchingplatform/study/repository/study/participant/ParticipantJpaRepository.java
@@ -2,7 +2,7 @@ package com.ludo.study.studymatchingplatform.study.repository.study.participant;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ludo.study.studymatchingplatform.study.domain.study.Participant;
+import com.ludo.study.studymatchingplatform.study.domain.study.participant.Participant;
 import com.ludo.study.studymatchingplatform.study.domain.id.ParticipantId;
 
 public interface ParticipantJpaRepository extends JpaRepository<Participant, ParticipantId> {
